tidy up mapper and remove redundant files
