+++ src/utils.ts
@@ -1,0 +1,5 @@
+import { XMLParser } from "fast-xml-parser";
+
+export const xmlParser = new XMLParser({
+  ignoreAttributes: false,
+});
